+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 2.4.9
+
+* generate Guid into SafeBehaviorSubject
+
 ### 2.4.8
 
 * add a condition to the Any Method